+++ src/ReinforcementLearningCore/src/policies/policies.jl
@@ -1,3 +1,4 @@
 include("agent.jl")
+include("agent_srt_cache.jl")
 include("random_policy.jl")
 include("q_based_policy.jl")